New function icompetition
For a given genotype g, returns a dictionary that maps every empty site
`es` adjecent to g to a tuple containing
a. indices of all neighbors of type g adj. to `es`
b. indices of all neighbors *not* of type g and not empty.

This is allows to measure the degree of competition for any empty site
between a specific genotype and every other.

Function is exported.